start of refactoring, simple dynamic relations
